seven segment code

+++ lib/spi.h
@@ -53,6 +53,6 @@ void spi_init(_SPI *self, uint16_t *SPIxSTAT, uint16_t *SPIxCON1,
               int16_t MOSIrpnum, int16_t SCKrpnum);
 void spi_open(_SPI *self, _PIN *MISO, _PIN *MOSI, _PIN *SCK, float freq);
 void spi_close(_SPI *self);
-uint8_t spi_transfer(_SPI *self, uint8_t val);
+uint16_t spi_transfer(_SPI *self, uint16_t val);
 
 #endif
